Change pip to user install to suppress errors
Without the --user argument, pip attempts to install to /usr/local/bin,
which it apparently cannot access without sudo (regardless of whether
the directory is owned by the current user or by root). With or without
the argument, the error does not seem to have any effect: installed
packages are still globally available for execution on the command line.