Remove Zsh dependency
Zsh was specified in the shebang because it provides a) easy color
support and b) the ability to run `source .zshrc` in the script.

Despite this, it does not make sense to require Zsh to run a script that
installs Zsh on fresh systems. With /bin/zsh in the shebang, this script
would not run on a fresh system.

It also is not necessary to source the .zshrc in this script: though
it is convenient, it is not too much to ask the user to run it
themselves or start a new login shell. The only functionality lost is
the ability to have environment variables and Zsh themes applied
immediately.

The message at the end of the script was provided to remind the user
that certain functionality will not be available until sourcing the
.zshrc or starting a new login shell.