Add default ESLint configuration
Without a project-level configuration or default configuration, Emacs
Flycheck complains about the lack of an ESLint configuration
file. This default configuration seeks to be as unopinionated as
possible by turning off all of the ESLint settings, so as to essentially
render it useless - the ideal for a project that either does not have a
linter configuration or prefers to use JSHint.