Fix bash profile exiting with 1
When `.locals` did not exist, the script would not return an error,
but would `exit 1`. This change now does not attempt to `source` the
file if it does not exist *and* exits with `0` whether the file exists
or not.

This fixes the prompt being red when it starts, indicating a
script that exited with 1, which is a bad way to start a terminal
session.