Change Zsh theme to 'simple' for ansi-term
`ansi-term` text modification has never worked correctly with the
`nicoulaj` theme due to the Git information on the end of each line of
the prompt. `agnoster` has all of its information on the left side of
the prompt, but tends to jump around when text is modified. Finally,
`simple` passes the litmus test:

1. With `ZSH_THEME` set to `simple`, write a command in `ansi-term`
   such as: `TEST=word source ~/.bashrc`
2. Switch to `term-line-mode` with `C-x C-j` and change `TEST` to
   `number`.
3. Switch back to `char-mode` with `C-x C-k` and hit enter to execute
   the command.

Throughout those steps, the current line should have remained in the
same position in the buffer. The executed command should have no extra
whitespace appended to the end of it (as can be examined by going back
into line mode and selecting directly after the command).