Fix startup error caused by code outside of mode hook
The org-mode verbatim string quotes fix was being run outside of any
mode hooks or anything, which is a problem when it's run at Emacs
start, before org-mode has been initialized, because it relies on
variables declared by org-mode.

This problem didn't turn up earlier because I had only tested it with
`eval-buffer` while Emacs was already running org-mode in a different
buffer.