Change meta keys from command to option on MacOS
This is more natural for the pinky finger to reach, because the option
key is right next to the control key, but the command key is two keys
over from it.

This also makes Emacs.app consistent with Emacs running in a terminal,
because the option key is usually the meta key in a terminal (or at
least iTerm). When I'm used to using Cmd for meta in Emacs.app and I
switch to Emacs in terminal, I often press Cmd + W instead of Opt + W,
closing the terminal window by accident.