Fix non-interactive login shell not initializing nodenv
VS Code executes tasks with a non-interactive login shell, like: `zsh  -l -c [command]`. If it were running bash, it would source `.bash_profile` (and not `.bashrc`) on a non-interactive login shell, but with Zsh, it only sources `.zprofile` on a non-interactive login shell (and `.zshrc` is sourced in only interactive shells, whether they are login or not). See [this answer](https://superuser.com/a/187673/201849) for more details.