Switch to Nodenv for Node and npm package installation
Nodenv can handle getting and installing the latest version of Node automatically
without relying on a package manager. It can also handle installing
a list of npm packages automatically. Installing npm packages this way
is better because the list of packages will be installed when *any* new
version of Node is installed in the future.