Remove grep
The Brew formula for grep makes the binary available as `g` by default,
which conflicts with the Zsh Git aliases. To make it install as `grep`
and overwrite the default system `grep`, one must pass
`--with-default-names` to `brew install grep`, but this is difficult to
do automatically with the package names specified in an array.
Since `grep` comes with macOS, it is not worth the added complexity to
refactor `brew.sh` to support this one situation.