Add function to get GitHub releases
This function can be called as `get-release` on the command line. The
function relies on the `hub` GitHub CLI.

If continuous integration is running when the function is called, it
will wait for it to finish, printing the status of each check every
five seconds. When it finishes, it will print only the releases that
were added instead of the whole list.