Change merge-pr to update base before merging
GitHub allows pull requests to be merged even when they are out of sync
with the base branch. This causes problems when using the `merge-pr`
function, because GitHub will reject the push if the branch is missing
commits from its remote counterpart. Pulling the latest changes into
the base branch before merging fixes the issue.