Switch to manual npm-merge-driver setup
`npm-merge-driver install` adds duplicate lines when customizations like
`-diff` are added to the `attributes` file. The `npm-merge-driver` setup
is not very complicated. Adding the necessary properties to the Git
config, checking in a Git attributes file, and modifying `git.sh` to put
the attributes file in place is all that was needed.