Get latest Node version correctly
Before, the script naively returned the last version in the list, which
was not always the latest version. For example, sometimes a new version
on an LTS branch would be released after the latest version on the
latest branch.

Sorting by version alone does not work because the versions are not
zero-padded. Sorting first by the module version and then by the version
works correctly.