Fix directory detection
The old "DIR" variable approach had a few issues:
- used capital letters, indicating an environment variable, instead of
lowercase
- used `$0` instead of `${BASH_SOURCE[0]}`, so it didn't work with
`source`
- since it was overwritten on each subsequent `source`, it didn't work
when files were in different directories
- because it was overwritten, it couldn't be declared as `readonly`

For dotfiles, "DIR" has been replaced with uniquely-named
`_dir_<filename>` variables, like `_dir_functions`. This is a
workaround and not a great long-term solution. These variables now
support `source` and are `readonly`.

For `lib/` scripts, "DIR" has been enhanced, but is still
overwritten. This works because the scripts are all in the same
directory, but is not a good long-term solution.

In the process, I discovered that I was sourcing Bash scripts from
Zsh scripts without much thought. It turns out that there were some
incompatibilities with `${BASH_SOURCE[0]}`, namely that it was not
set and even if it were set, Zsh indexes arrays starting at one
instead of at zero. Sourcing with the `emulate` command and manually
setting `BASH_SOURCE` resolves the issue, but is not a very clean
solution.