Prepend Python bin directory instead of appending
This was causing `PATH` ordering issues. This directory happens to be
`~/.local/bin` on Linux. The `append_path` call that adds this
directory to `PATH` was earlier in `.exports` than the `prepend_path`
call that explicitly adds `~/.local/bin` to `PATH`, so the latter
call wasn't running at all. This resulted in `~/.local/bin` at the end
of `PATH` instead of the beginning, so binaries in that directory were
not overriding binaries in other directories.