Switch from i3status to i3blocks
i3blocks offers greater control over the content of the statusline
without having to write a wrapper script. The `status-pulseaudio`
script introduced here is _almost_ equivalent to the i3status volume
output, except that it doesn't report the mute status of the sink. It
includes the enhancement of displaying a headphones/speaker icon
depending on which sink is the current default.