Avoid shadowing Mercurial executable
I chose `hg` as the name of my wrapper for ripgrep, `rg`, that enables
Kitty's `hyperlinked_grep` kitten when supported. I hadn't realized
that `hg` is also the executable name for the Mercurial version
control CLI. This conflict manifested in `jest --watch` hanging
indefinitely on the message "Determining test suites to run..." as it
ran `rg root` (originally `hg root`, before the `hg` script ran `exec
rg root`).

The execution chain in Jest, for reference:
- `jest --watch`
- `getChangedFilesPromise` (core Jest)
- `getChangedFilesForRoots` (`jest-changed-files` from here on out)
- `findRepos`
- `hg.getRoot` runs `hg root`, which returns dozens of lines of random
text instead of the directory path expected by Jest
- `hg.findChangedFiles` runs shell command `hg status` once per
line (unclear if this was actually happening or not)